refactor: use database structure in laravel
